move functions for sequences of items into payloadsettings
