+++ src/firestarter/Environment/X86/Payload/AVXPayload.cpp
@@ -35,8 +35,9 @@ auto AVXPayload::compilePayload(const environment::payload::PayloadSettings& Set
 
   // Compute the sequence of instruction groups and the number of its repetions
   // to reach the desired size
-  auto Sequence = generateSequence(Settings.instructionGroups());
-  auto Repetitions = getNumberOfSequenceRepetitions(Sequence, Settings.linesPerThread());
+  auto Sequence = Settings.sequence();
+  auto Repetitions =
+      environment::payload::PayloadSettings::getNumberOfSequenceRepetitions(Sequence, Settings.linesPerThread());
 
   // compute count of flops and memory access for performance report
   environment::payload::PayloadStats Stats;
@@ -73,9 +74,12 @@ auto AVXPayload::compilePayload(const environment::payload::PayloadSettings& Set
   const auto RamSize = Settings.ramBufferSizePerThread();
 
   // calculate the reset counters for the buffers
-  const auto L2LoopCount = getL2LoopCount(Sequence, Settings.linesPerThread(), L2Size);
-  const auto L3LoopCount = getL3LoopCount(Sequence, Settings.linesPerThread(), L3Size);
-  const auto RamLoopCount = getRAMLoopCount(Sequence, Settings.linesPerThread(), RamSize);
+  const auto L2LoopCount =
+      environment::payload::PayloadSettings::getL2LoopCount(Sequence, Settings.linesPerThread(), L2Size);
+  const auto L3LoopCount =
+      environment::payload::PayloadSettings::getL3LoopCount(Sequence, Settings.linesPerThread(), L3Size);
+  const auto RamLoopCount =
+      environment::payload::PayloadSettings::getRAMLoopCount(Sequence, Settings.linesPerThread(), RamSize);
 
   asmjit::CodeHolder Code;
   Code.init(asmjit::Environment::host());
@@ -329,7 +333,7 @@ auto AVXPayload::compilePayload(const environment::payload::PayloadSettings& Set
     }
   }
 
-  if (getRAMSequenceCount(Sequence) > 0) {
+  if (environment::payload::PayloadSettings::getRAMSequenceCount(Sequence) > 0) {
     // reset RAM counter
     auto NoRamReset = Cb.newLabel();
 
@@ -342,7 +346,7 @@ auto AVXPayload::compilePayload(const environment::payload::PayloadSettings& Set
     // adds always two instruction
     Stats.Instructions += 2;
   }
-  if (getL2SequenceCount(Sequence) > 0) {
+  if (environment::payload::PayloadSettings::getL2SequenceCount(Sequence) > 0) {
     // reset L2-Cache counter
     auto NoL2Reset = Cb.newLabel();
 
@@ -355,7 +359,7 @@ auto AVXPayload::compilePayload(const environment::payload::PayloadSettings& Set
     // adds always two instruction
     Stats.Instructions += 2;
   }
-  if (getL3SequenceCount(Sequence) > 0) {
+  if (environment::payload::PayloadSettings::getL3SequenceCount(Sequence) > 0) {
     // reset L3-Cache counter
     auto NoL3Reset = Cb.newLabel();
