+++ src/firestarter/Environment/X86/Payload/ZENFMAPayload.cpp
@@ -34,8 +34,9 @@ auto ZENFMAPayload::compilePayload(const environment::payload::PayloadSettings& 
 
   // Compute the sequence of instruction groups and the number of its repetions
   // to reach the desired size
-  auto Sequence = generateSequence(Settings.instructionGroups());
-  auto Repetitions = getNumberOfSequenceRepetitions(Sequence, Settings.linesPerThread());
+  auto Sequence = Settings.sequence();
+  auto Repetitions =
+      environment::payload::PayloadSettings::getNumberOfSequenceRepetitions(Sequence, Settings.linesPerThread());
 
   // compute count of flops and memory access for performance report
   environment::payload::PayloadStats Stats;
@@ -72,9 +73,12 @@ auto ZENFMAPayload::compilePayload(const environment::payload::PayloadSettings& 
   const auto RamSize = Settings.ramBufferSizePerThread();
 
   // calculate the reset counters for the buffers
-  const auto L2LoopCount = getL2LoopCount(Sequence, Settings.linesPerThread(), L2Size);
-  const auto L3LoopCount = getL3LoopCount(Sequence, Settings.linesPerThread(), L3Size);
-  const auto RamLoopCount = getRAMLoopCount(Sequence, Settings.linesPerThread(), RamSize);
+  const auto L2LoopCount =
+      environment::payload::PayloadSettings::getL2LoopCount(Sequence, Settings.linesPerThread(), L2Size);
+  const auto L3LoopCount =
+      environment::payload::PayloadSettings::getL3LoopCount(Sequence, Settings.linesPerThread(), L3Size);
+  const auto RamLoopCount =
+      environment::payload::PayloadSettings::getRAMLoopCount(Sequence, Settings.linesPerThread(), RamSize);
 
   asmjit::CodeHolder Code;
   Code.init(asmjit::Environment::host());
@@ -286,7 +290,7 @@ auto ZENFMAPayload::compilePayload(const environment::payload::PayloadSettings& 
   }
 
   Cb.movq(TempReg, IterReg); // restore iteration counter
-  if (getRAMSequenceCount(Sequence) > 0) {
+  if (environment::payload::PayloadSettings::getRAMSequenceCount(Sequence) > 0) {
     // reset RAM counter
     auto NoRamReset = Cb.newLabel();
 
@@ -300,7 +304,7 @@ auto ZENFMAPayload::compilePayload(const environment::payload::PayloadSettings& 
     Stats.Instructions += 2;
   }
   Cb.inc(TempReg); // increment iteration counter
-  if (getL2SequenceCount(Sequence) > 0) {
+  if (environment::payload::PayloadSettings::getL2SequenceCount(Sequence) > 0) {
     // reset L2-Cache counter
     auto NoL2Reset = Cb.newLabel();
 
@@ -314,7 +318,7 @@ auto ZENFMAPayload::compilePayload(const environment::payload::PayloadSettings& 
     Stats.Instructions += 2;
   }
   Cb.movq(IterReg, TempReg); // store iteration counter
-  if (getL3SequenceCount(Sequence) > 0) {
+  if (environment::payload::PayloadSettings::getL3SequenceCount(Sequence) > 0) {
     // reset L3-Cache counter
     auto NoL3Reset = Cb.newLabel();
